Struct wasmer_engine_universal_artifact::SerializableModule

pub struct SerializableModule {
    pub compilation: SerializableCompilation,
    pub compile_info: CompileModuleInfo,
    pub data_initializers: Box<[OwnedDataInitializer]>,
    pub cpu_features: u64,
}

Serializable struct that is able to serialize from and to a UniversalArtifactInfo.

Fields
compilation: SerializableCompilation

The main serializable compilation object

compile_info: CompileModuleInfo

Compilation informations

data_initializers: Box<[OwnedDataInitializer]>

Datas initializers

cpu_features: u64

CPU Feature flags for this compilation

Implementations

impl SerializableModule

pub fn serialize(&self) -> Result<Vec<u8>, SerializeError>

Serialize a Module into bytes The bytes will have the following format: RKYV serialization (any length) + POS (8 bytes)

pub unsafe fn deserialize(
metadata_slice: &[u8]
) -> Result<Self, DeserializeError>

Deserialize a Module from a slice. The slice must have the following format: RKYV serialization (any length) + POS (8 bytes)

Safety

This method is unsafe since it deserializes data directly from memory. Right now we are not doing any extra work for validation, but rkyv has an option to do bytecheck on the serialized data before serializing (via rkyv::check_archived_value).

pub fn deserialize_from_archive(
archived: &ArchivedSerializableModule
) -> Result<Self, DeserializeError>

Deserialize a compilation module from an archive