Trait wasmer_engine_universal_artifact::ArtifactCreate

pub trait ArtifactCreate: Send + Sync + Upcastable + MemoryUsage {
    fn module(&self) -> Arc<ModuleInfo>;

    fn module_ref(&self) -> &ModuleInfo;

    fn module_mut(&mut self) -> Option<&mut ModuleInfo>;

    fn features(&self) -> &Features;

    fn cpu_features(&self) -> EnumSet<CpuFeature>;

    fn memory_styles(&self) -> &PrimaryMap<MemoryIndex, MemoryStyle>;

    fn table_styles(&self) -> &PrimaryMap<TableIndex, TableStyle>;

    fn data_initializers(&self) -> &[OwnedDataInitializer];

    fn serialize(&self) -> Result<Vec<u8, Global>, SerializeError>;

    fn serialize_to_file(&self, path: &Path) -> Result<(), SerializeError> { ... }
}

An Artifact is the product that the Engine implementation produce and use.

The Artifact contains the compiled data for a given module as well as extra information needed to run the module at runtime, such as ModuleInfo and Features.

Required methods

fn module(&self) -> Arc<ModuleInfo>

Return a reference-counted pointer to the module

fn module_ref(&self) -> &ModuleInfo

Return a pointer to a module.

fn module_mut(&mut self) -> Option<&mut ModuleInfo>

Gets a mutable reference to the info.

Note: this will return None if the module is already instantiated.

fn features(&self) -> &Features

Returns the features for this Artifact

fn cpu_features(&self) -> EnumSet<CpuFeature>

Returns the CPU features for this Artifact

fn memory_styles(&self) -> &PrimaryMap<MemoryIndex, MemoryStyle>

Returns the memory styles associated with this Artifact.

fn table_styles(&self) -> &PrimaryMap<TableIndex, TableStyle>

Returns the table plans associated with this Artifact.

fn data_initializers(&self) -> &[OwnedDataInitializer]

Returns data initializers to pass to InstanceHandle::initialize

fn serialize(&self) -> Result<Vec<u8, Global>, SerializeError>

Serializes an artifact into bytes

Provided methods

fn serialize_to_file(&self, path: &Path) -> Result<(), SerializeError>

Serializes an artifact into a file path

Implementations

impl dyn ArtifactCreate + 'static

pub fn downcast_ref<T>(&self) -> Option<&T> where
T: 'static,

Try to downcast the artifact into a given type.

pub fn downcast_mut<T>(&mut self) -> Option<&mut T> where
T: 'static,

Try to downcast the artifact into a given type mutably.