Module wasmer_engine_universal_artifact::trampoline

Trampolines for libcalls.

This is needed because the target of libcall relocations are not reachable through normal branch instructions.

Constants

AARCH64_TRAMPOLINE

X86_64_TRAMPOLINE

Functions

get_libcall_trampoline

Returns the address of a trampoline in the libcall trampolines section.

libcall_trampoline_len

Returns the length of a libcall trampoline.

make_libcall_trampolines

Creates a custom section containing the libcall trampolines.

make_trampoline